requirements.txt
asgiref==3.3.0
autopep8==1.5.4
bootstrap4==0.1.0
certifi==2020.11.8
cffi==1.14.4
chardet==4.0.0
cloudinary==1.23.0
codecov==2.1.11
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
coveralls==2.2.0
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap3==14.2.0
django-braces==1.14.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-crispy-forms==1.10.0
django-filter==1.1.0
django-generic-aggregation==0.4.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-readonly-field==1.0.5
django-registration==3.1.1
django-rest-knox==4.1.0
django-rql==3.8.1
django-simple-ratings==0.3.3
django-star-ratings==0.9.1
django-tinymce==3.1.0
djangorestframework==3.12.2
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
<<<<<<< HEAD
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.8
=======
lark-parser==0.8.2
>>>>>>> main
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-coveralls==2.9.3
python-decouple==3.3
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
<<<<<<< HEAD
travis==0.0.3
uritemplate==3.0.1
=======
toml==0.10.2
>>>>>>> main
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
